requirements.txt
paramiko
python-novaclient
